feat(pswp-v5): split stories into several files
